requirements.txt
streamlit
joblib
pyyaml
numpy
scikit-learn
sentence-transformers
openai
